app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.1",
    "cryptography>=44.0.2",
    "dotenv>=0.9.9",
    "geopy>=2.4.1",
    "httpx>=0.28.1",
    "piexif>=1.1.3",
    "pillow>=11.1.0",
    "psycopg2==2.9.10",
    "pyheif>=0.8.0",
    "python-json-logger>=3.3.0",
    "python-telegram-bot[job-queue]>=22.0",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.40",
]
